chore: move cardstyles
