+++ frontend/components/AgentWallet/Withdraw/WithdrawFromAgentWallet.tsx
@@ -1,5 +1,7 @@
 import { Button, Flex, Input, Modal, Typography } from 'antd';
-import { CSSProperties, useCallback, useState } from 'react';
+import { useCallback, useState } from 'react';
+
+import { cardStyles } from '@/components/ui/cardStyles';
 
 import { ChainAndAmountOverview } from './ChainAndAmountOverview';
 import { useWithdrawFunds } from './useWithdrawFunds';
@@ -10,11 +12,6 @@ import {
 } from './WithdrawStatusMessage';
 
 const { Text } = Typography;
-
-export const cardStyles: CSSProperties = {
-  width: 552,
-  margin: '0 auto',
-} as const;
 
 const PasswordLabel = () => (
   <Text className="text-sm text-neutral-tertiary">
