+++ frontend/components/PearlWallet/Withdraw/SelectAmountToWithdraw/SelectAmountToWithdraw.tsx
@@ -2,9 +2,10 @@ import { Button, Flex, Typography } from 'antd';
 
 import { BackButton } from '@/components/ui/BackButton';
 import { CardFlex } from '@/components/ui/CardFlex';
+import { cardStyles } from '@/components/ui/cardStyles';
 
 import { usePearlWallet } from '../../PearlWalletProvider';
-import { cardStyles, PearlWalletToExternalWallet } from '../common';
+import { PearlWalletToExternalWallet } from '../common';
 import { TokenAmountInput } from './TokenAmountInput';
 
 const { Title } = Typography;
