+++ frontend/components/ui/cardStyles.ts
@@ -1,0 +1,6 @@
+import { CSSProperties } from 'styled-components';
+
+export const cardStyles: CSSProperties = {
+  width: 552,
+  margin: '0 auto',
+} as const;
